feat(currency): integrate exchange rate API with Redis caching and price conversion

+++ src/main/java/com/commercial/Pont/Commercial/dtos/requestDtos/CreateMyAnnonceRequestDto.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Set;
@@ -26,7 +27,7 @@ public class CreateMyAnnonceRequestDto {
     private String certification;
     private String description;
     private AnnouncementType type;
-    private Double prix;
+    private BigDecimal prix;
     private String devise;
     private Double quantite;
     private String uniteQuantite;
